requirements.txt
anthropic
rich
tavily-python
ollama
groq
pypandoc
fastapi
uvicorn
twilio
python-decouple
sqlalchemy
psycopg2-binary
python-multipart
pyngrok
flask
Jinja2
MarkupSafe
PyJWT
Werkzeug
certifi
charset-normalizer
click 
flask 
idna 
importlib-metadata 
itsdangerous 
pytz
requests
twilio
urllib3
zipp
google-cloud
google-cloud-storages
langchain
langchain_community